done, it is working fine

+++ Main/Main.cpp
@@ -1,6 +1,7 @@
 #include "../Headers/Genome.hpp"
+#include <mpi.h>
 
-int main ()
+int main (int argc, char **argv)
 {
     //ifstream file("./input/parameters.csv"); // Replace with your CSV file path
     //string line;
@@ -16,17 +17,19 @@ int main ()
 
     //ifstream Predifined ("./input/parameters.csv");
     int N = 20;
-    int st = 1000;
+    int st = 2000;
     int Nn_net = 10000;
     
     
     lineCount = 1;
-    float env_reF[] = {0.0 , 0.2, 0.5};
+    float env_reF[] = {0.0 , 0.2, 0.5, 0.8};
 
     string comnd = "rm -rf ./ARCHIVE";
     system (comnd.c_str());
     comnd = "mkdir ./ARCHIVE";
     system (comnd.c_str());
+
+    
 
     //for (int jk=0 ; jk<lineCount ; jk++)
     //{
@@ -35,56 +38,75 @@ int main ()
             //envi_func = a;
             int envi_func = 0;
             
-
-            for (int L=0 ; L<3 ; L++)
+            int world_size, world_rank;
+            MPI_Init(&argc, &argv);
+            MPI_Comm_size(MPI_COMM_WORLD, &world_size);
+            MPI_Comm_rank(MPI_COMM_WORLD, &world_rank);
+            
+            for (int L = 0 ; L < 4 ; L++) 
             {
-                float  env_ref = env_reF[L];
+                float env_ref = env_reF[L];
                 float Mute_R = 0.001;
                 envi_func = 0;
-                int number_runs = 4;
-                string rem = "rm -r ./Outputs/*";
-                system (rem.c_str());
+                int number_runs = 6;
+                
+                if (world_rank == 0) 
+                {
+                    // Assuming env_reF is available here...
+                    string rem = "rm -r ./Outputs/*";
+                    system (rem.c_str());
+                }
+                
+                MPI_Barrier(MPI_COMM_WORLD);
 
-                for (int RUN=0 ; RUN < number_runs ; RUN++)
-                {   
+                for (int RUN = world_rank; RUN < number_runs; RUN += world_size) 
+                {
                     genome GM;
                     GM.Setting_initial_values(N, st, Nn_net, Mute_R, env_ref, envi_func, RUN);
                 }
                 
-                string add = "st_" + to_string(st) + "," + "nn_" + to_string(Nn_net) + "," + "mu_" + to_string(Mute_R) + "," + "ref_env_=" + to_string(env_ref);
-                switch (envi_func)
+                MPI_Barrier(MPI_COMM_WORLD);
+                
+                if (world_rank == 0)
                 {
-                    case 0:
-                    comnd = "mkdir ./ARCHIVE/Linear_input/";
-                    system (comnd.c_str());
-                    add = "./ARCHIVE/Linear_input/" + add + "/";
-                    break;
+                    string add = "st_" + to_string(st) + "," + "nn_" + to_string(Nn_net) + "," + "mu_" + to_string(Mute_R) + "," + "ref_env_=" + to_string(env_ref);
+                    
+                    switch (envi_func)
+                    {
+                        case 0:
+                        comnd = "mkdir ./ARCHIVE/Linear_input/";
+                        system (comnd.c_str());
+                        add = "./ARCHIVE/Linear_input/" + add + "/";
+                        break;
 
-                    case 1:    
-                    comnd = "mkdir ./ARCHIVE/Gaus_input/";
-                    system (comnd.c_str());
-                    add = "./ARCHIVE/Gaus_input/" + add + "/";
-                    break;
-                
-                    case 2:
-                    comnd = "mkdir ./ARCHIVE/Step_input/";
-                    system (comnd.c_str());
-                    add = "./ARCHIVE/Step_input/" + add + "/";
-                    break;
+                        case 1:    
+                        comnd = "mkdir ./ARCHIVE/Gaus_input/";
+                        system (comnd.c_str());
+                        add = "./ARCHIVE/Gaus_input/" + add + "/";
+                        break;
+                    
+                        case 2:
+                        comnd = "mkdir ./ARCHIVE/Step_input/";
+                        system (comnd.c_str());
+                        add = "./ARCHIVE/Step_input/" + add + "/";
+                        break;
 
-                    case 3:
-                    comnd = "mkdir ./ARCHIVE/Negative_input/";
+                        case 3:
+                        comnd = "mkdir ./ARCHIVE/Negative_input/";
+                        system (comnd.c_str());
+                        add = "./ARCHIVE/Negative_input/" + add + "/";
+                        break;
+                    }
+
+                    comnd = "mkdir " + add;
                     system (comnd.c_str());
-                    add = "./ARCHIVE/Negative_input/" + add + "/";
-                    break;
+                    
+                    comnd = "cp -r ./Outputs/ " + add;
+                    system (comnd.c_str());
                 }
-
-                comnd = "mkdir " + add;
-                system (comnd.c_str());
                 
-                comnd = "cp -r ./Outputs/ " + add;
-                system (comnd.c_str());
             }
+            MPI_Finalize();
         //}
     //}
 }
